+++ src/collector/main.py
@@ -4,8 +4,8 @@ from datetime import datetime
 
 from src import META_DATA_PATH, ABI_FILEPATH, SNAPSHOT_FILE_PATH
 from src.collector.db.ops import create_tables, get_schain_stats
-from src.collector.endpoints import get_all_names, is_dkg_passed, get_schain_endpoint
-from src.collector.fetchers.statistics import Collector, PricesCollector
+from src.collector.core.endpoints import get_all_names, is_dkg_passed, get_schain_endpoint
+from src.collector.core.fetchers import Collector, PricesCollector
 from src.utils.helper import daemon, write_json
 from src.utils.logger import init_logger
 from src.utils.meta import create_meta_file, get_meta_file, update_meta_file
